making health double

+++ 2ndTermProject/AlienArmy/AlienDrone.h
@@ -4,7 +4,7 @@
 class AlienDrone : public ArmyUnit
 {
 public:
-	AlienDrone(Game* p, int HP, int pow, int ID_, int cap, int Tj, Unit U = AD);
+	AlienDrone(Game* p, double HP, double pow, int ID_, int cap, int Tj, Unit U = AD);
 	friend ostream& operator<<(ostream& COUT, AlienDrone* PAssed_AD);
 	virtual bool Attack();
 };
